Restructure CI variables modal help links
In the CI/CD variable modal, we had a lot of
icons and link to help user learn more. With this
change, we centralize all "flags" under one icon
and link next to the Flags label and remove
individual icons on each flag. We also change
the size of the icons and align a few things.

Changelog: changed