Add confirmation modal when retrying a bridge
We now have the ability to retry a trigger job behind a
FF `ci_recreate_downstream_pipeline`. This commit adds a
confirmation message to the modal to ensure users are not
surprised when it spawns a whole new pipeline. There is also
the ability to not show the modal again.

Changelog: added