Skip zoekt RSpec tests in CI in more places
When we added these tests in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/105049 we intended
to skip them everywhere. Now I noticed I actually missed a couple of
rspec commands where they should also be skipped. I was surprised when I
saw the fail-fast job in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/110479/pipelines
fail so I spotted that I'd missed this in `fail-fast`. So I grep'd and
found a few more places we skip quarantine so I added it to all of those
too.