Filter latest downstream pipelines in pipeline editor
Retrying a trigger job will create a new downstream pipeline. Which
means the pipeline graphs can show multiple outdated downstream
pipelines.

This MR creates a util function for filtering out the outdated
downstream pipelines via the `sourceJob.retried` property. If the the
source job (trigger job) has `retried = false`, this means that it is
the latest trigger job, which means the pipeline associated to it will
be rendered.

This is tested on the mini pipeline graph for the pipeline editor, which
uses GraphQL. Updates for other mini pipeline graphs (some of which use
REST or Rails injection) will be done in future MRs.

Since the retry trigger job feature is not yet available, this should
not affect how we render existing pipeline graphs today.