Ensures we don't create duplicate records
We are seeing errors for attempts to create Ci::PipelineArtifact records
for Code Quality MR diffs that violate the uniqueness constraint on that
table. The return statement that checks for
pipeline.has_codequality_mr_diff_report? in the execute statement should
be enough to prevent this, but the error may be coming from a race
condition. This change will prevent the error by deleting any existing
records before creating a new one.

Changelog: fixed