Revert "Extend gitlab:gitaly:check task with readiness verification"
This reverts commit 5923d4d245c60049dbb97e799115c7e05ad67284.

The Gitaly readiness check is invoked by a GitLab Rails readiness check
endpoint `GET /-/readiness?all=1` and the `gitlab:gitaly:check` rake
task. The `ReadinessCheck` RPC, as currently implemented, is not fit for
use by the Rails API as it was originally intended as a diagnostic tool
and not something to be invoked frequently. The checks in the RPC open
multiple database pools and run heavy queries, which can consume
significant quantities of system resources. The rake task also has an
inherent issue with this RPC because the task calls Praefect through the
load balancer. This means that results could vary depending on which
Praefect node recieves the request.

To avoid these problems, this change removes the invocation of the
`ReadnessCheck` RPC. Future updates to this RPC will allow for the check
to be reintroduced.