Always convert dotenv files to UTF-8
Previously if Windows `powershell` were used with the `echo` command,
the dotenv artifact would be created in UTF-16LE format with the BOM
at the start of the file. When GitLab attempted to parse this file, it
would return a 500 error because the parser did not know how to handle
the invalid encoding.

To avoid this, before parsing the dotenv file, we force the encoding
to UTF-8 with the charlock_holmes gem and strip the BOM from the file.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/389995

Changelog: fixed