Add support for SVG rendering in design management
SVGs without dimension attributes were spanning in the design list.
In actual, height is not need on image but on parent div.
The parent div is of gl-intersection-observer.
As both objects are in flex container, apply flex grow
on the parent div to help image render properly.
Also, the offset dimensions are zero thus dimension calculation
was breaking. If offset dimensions are zero, calculate the image base
dimensions from natural dimensions. Set the dimension on mount so that
zoom works properly. Add rspec to check if the SVG without dimensions
is loading properly. Update the snapshot for the spec.

Changelog: fixed