Handle client disconnects better in workhorse
When a client disconnects before it has received a response, workhorse currently detects this as a 502 (logged and put into metrics).  However this ends up in metrics which for gitlab.com and GitLab Dedicated is counted as an error against the SLA, when it is entirely a client-side problem.  Nginx logs this as a 499 (a non-standard code, but a useful one).

This patch extends that behavior to Workhorse, such that when a client disconnects early, it is counted as a 499 and not a server error, because it is strictly the client's problem (and probably just a browser window being closed or similar)