Refactor specs: mock Gitlab::Database::LockWritesManager
The specs for Gitlab::Database::TablesLocker did verify that the
tables are actually locked by trying to interact with the tables.

However, the Gitlab::Database::TablesLocker is using the LockManager
for locking and unlocking the tables. The TableLocker responsibility
is figuring out what to lock and unlock.

By mocking the LockManager, we can make these specs run faster