Clear existing release on tag name change
When users went to select a tag that was in use and then later change it
to a tag not in use or create a new tag the error was not clearing.

Instead, once the tag name changes or we fail to fetch tag information,
we should clear out the existing release, as it is out of date.

Changelog: fixed