Highlight the difference between push and pull mirror API
**Problem**

It's easy to assume that remote mirrors API includes both push and pull
mirrors. It can cause a confusion why configured pull mirrors are not
visible in the response.

**Solution**

We mention pull mirrors on the documentation page, but it's not very
visible.

We can add a warning message that guides users how to configure pull
mirrors.