Introduce a class for generating a list of tables to lock
This is a refactoring. Locking tables is a two-step process:
- Find out which tables to lock
- Lock (or unlock) them

The Lock/Unlock logic is in the `LockWritesManager` class. But the
'which tables to lock' logic lived in the rake task. By moving this
logic to a new `TableLocker` class, we better separate responsibilties

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/384905