Fix public_image_exist? template spec helper
We to explicitly tell the server which type of manifests it is
looking for.

This is causing failures on
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/102893, because
the updated image uses a different manifest format. The response error
message is "OCI index found, but accept header does not support OCI
indexes".

Also throws errors on unexpected response statuses.