Filter out duplicated downstreams in pipeline graph
When the FF ci_recreate_downstream_pipeline is enabled, we need
to filter out downstream pipelines which source job has been retried.
This is to avoid seeing multiple downstreams related to a single
trigger job as these pipelines are no longer the "latest". When the
FF is off, it should not filter anything out and the behaviour should
stay the same.

Changelog: fixed