Add new `Terraform-Module.gitlab-ci.yml` CI/CD template
This change set introduces a new CI/CD template named
`Terraform-Module.gitlab-ci.yml` which can be used to easily deploy a
Terraform Module to the GitLab Terraform Registry.

Prior to this change the documentation contained a pipeline snippet
which could be used to achieve the same. This is a little cumbersome for
users though and it would be nice if that snippet would just exist as a
template - that's what this MR does ;)

The template comes as a *pipeline* and *job* template which makes it
easy to adapt to more complex use-cases like deploying multiple
Terraform Modules from the same pipeline.

This change set also contains the relevant new template metrics and
specs for the new templates.

Changelog: added
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/110493