Handle both "normal" discussions and image diff discussions
Image diff discussions don't use a table row, which causes
DOMPurify to strip the outer table since the child is a <div>.
This causes the .firstChild selector to fail on the result of
.querySelector( "table" ), since that always returns null.

An alternative solution here is to ONLY modify line that
selects the element to append, by modifying the expression
based on whether the note is `on_image` or not:

```
internalNote = sanitize(`<table>${noteEntity.diff_discussion_html}</table>`, {
    RETURN_DOM: true,
});

discussionElement.insertAdjacentElement( 'afterbegin', noteEntity.on_image ? internalNote.firstChild : internalNote.querySelector( "table" ).firstChild );
```

However, this is pretty "clever" (in a bad way) because it depends on the
reader knowing that DOMPurify will magically strip the <table> element if
its children aren't <tr>s. The if/else here makes it clear that there are
two potential parsing modes: one for image comments, and one for other comments.