gitaly_client: Always enable eager housekeeping for manual jobs
In 467b047c9d6b (gitaly_client: Wire up eager housekeeping strategy for
OptimizeRepository, 2023-01-30) we have introduced the logic to run
eager housekeeping tasks when an administrator explicitly requests them
to run. This is to address an issue where the heuristical housekeeping
strategy would not perform any optimizations if the repository is found
to already be well-optimized by Gitaly. This has been causing confusion
for administrators.

The change was rolled out to production systems successfully. Remove the
feature flag to unconditionally enable eager houskeeeping in the context
fo manual jobs.

Changelog: changed