Fix false positive in zoekt/searchable_repository_spec.rb
This test was for private projects was meant to be testing the logic of
`def use_zoekt?` in `ee/app/models/ee/project.rb`. The problem is that
the project was not in an indexed namespace so it was going to be false
already even if it was public or even if we comment out the logic for
checking if it's public.

This change makes it share the same namespace as `project` which is
in an indexed namespace (due to calling `zoekt_ensure_project_indexed`)
and thus this ensures the test actually validates the logic in
`def use_zoekt?` works as expected.