Add spec to test preloader in audi events report service
It was mentioned in a [separate MR][0] that the preloader was not
being tested.

This commit adds a spec that does that.

If the preloader lines is removed, all of the existing specs still
pass however the new spec fails with the following message:

```sh
Failure/Error: expect { described_class.new(params).csv_data.to_a }.not_to exceed_query_limit(control)

  Expected a maximum of 1 queries, got 2:

  Query Diff:
  -----------
  SELECT "users"."id", "users"."name", "users"."username" FROM "users"...
  -- (expected: 0, got: 1)
     WHERE "users"."id" IN (16, 18) ORDER BY "users"."id" ASC LIMIT 1000
```

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/110846#note_1264883203