Let users delete rules from protected environments
Only deployer rules for now, as approval rules come later.

The API is very explicit about only accepting one of group_id, user_id,
or access_level (even if the others are set to null), so we need to
filter out and find only properties that have a truthy value.