Add more logging to MergeService
While investigating https://gitlab.com/gitlab-org/gitlab/-/issues/389044,
it is noticeable that `MergeTrains::RefreshWorker` that utilizes
`MergeRequests::MergeService` stops logging after Git merge is
done.

A theory is the ensure block in `#commit` is stopping the
execution for some unknown reason yet.

This adds logging to that block to see if that block gets executed
and failing. It could potentially help in investigating the bug.