Fix depth check when adding existing epics
Before this change, when adding an existing epic to the hierarchy, we
did not calculate the right depth during validation and didn't allow
creating hierarchies that are 7 levels deep.

For new epics it worked as we're not accounting for the Epic's
hierarchy.

This off-by one error is now fixed as the new calculation uses `1` as
the minimum hierarchy.

Changelog: fixed
EE: true