Update GraphQL API for work item quick actions
Quick action commands can be included in
the description change of a work item
when updated through the GraphQL API.

The mutation resolver for work item update
will run QuickActions::InterpretService and pass
the parsed update params such as 'assignee_ids'
for the '/assigne' quick action to the underlying widgets.

We are also updating the create/update services for issues
so they do NOT handle any quick action updates intended
for work items.

Changelog: added