Add a table for design repo and a class for the corresponding git repo
This commit re-factors the class DesignManagement::Repository such that
it now represents a project's design repo on the whole, while a new
class, DesignManagement::GitRepository is a wrapper for the
corresponding GitLab managed git repo.

The calls to the git repo stay the same, because the Repository class
simply delegates the method calls to
`DesignManagement::Repository.git_repo`.

DesignManagement::Repository, is now backed by a table, associating the
design repo to the project is belongs to.

This change is introduced in preparation for
https://gitlab.com/gitlab-org/gitlab/-/issues/355660

Changelog: added