Refactor CI includes
This is mostly about redundant controls.
- Removed "next unless artifact_job" from "Artifact#content" because
this method will not be called if it is not valid.
- Removed "next unless creating_child_pipeline?" from "Artifact#content"
because this method will not be called if it is not valid.
- Replaced "return unless context.project" with "return unless valid?"
in the "Local#masked_*" methods because we don't need those methods
unless valid.
- Removed "next unless can_access_local_content?" and
"return unless sha" from "Project#fetch_local_content" because
this method will not be called if it is not valid.
- Removed "next unless project" from "Project#sha" because
this method will not be called if it is not valid.
- Replaced "return unless project" with "valid?" in the
"Project#masked_*" methods because we don't need those methods
unless valid.