Fix CI runner validation
When the runner is not persisted yet, and the runner_namespaces relation
is not loaded, and we call runner_namespaces.one?, it tries to do a
count query which ends up being 0 because there is no runner_id yet.

This leads to runner_namespaces.one? returning false even though there
is exactly one group to be inserted as the namespace.