Fetch latest license scanning build from scanner
The license compliance page generates a link to
the last successful license scanning job. This
job should change depending on the source of the
licenses found - SBoM vs a license scanning report.
Moving the logic for this to the license scanning
scanner makes it possible to change which build is
returned based on the status of the FF that
toggles the SBoM scanner. An initial implementation
of the latest build method is added to the SBoM
scanner and can be changed as requirements evolve.