Revert kerrizor/add-additional-check-to-rescue-search-timeouts
Reverts https://gitlab.com/gitlab-org/gitlab/-/merge_requests/110672,
which added a guard testing for @search_timeout_occurred. This guard was
added to prevent nav elements on the page from attempting to re-run the
failing search query for populating state counts, but this guard is made
obsolete by https://gitlab.com/gitlab-org/gitlab/-/merge_requests/111084
which we added in a later MR. The guard itself is causing a small
regression, in that it hides the bulk MR edit button from displaying
correctly.

Changelog: fixed