Validate notification integrations channel limit
This change adds a new validation to our Slack and Mattermost
notification integrations that each event trigger is configured for at
most 10 unique channels.

It also adds a tweak so we send to a unique list of channels, to avoid
the ability for the integrations to post the same thing to the same
channel.

Changelog: changed