Mark diff tree entries as loaded when we place them into the store
SET_DIFF_DATA_BATCH already assumes we may not have the full list
of files. So whether we're loading in a batch of files, or one
single file, this mutation can handle the update.

This new code updates each tree entry that matches one of the
inbound diff files to read as `diffLoaded = true`. By default
this property is `false`.