Release improved cache headers for archive and raw blobs
This feature flag has been introduced as a safety measure for
enabling new cache headers on Sass. It has been enabled for
a while and now can be removed. There is no need in following
the process of enabling it by default and then removing it all
along.

Changelog: performance