Fallback to SBOM license scanning
When the license scanning sbom scanner FF
is enabled, the license_scanning artifacts
should be prioritized over the SBOM reports.
When finding the scanner for a project, we
determine this by returning the scanner that
has the most recent data. For pipelines, we
first check if it has a license_scanning report
and then fallback to using an SBoM scanner.