Serialize panel data for Super Sidebar
This adds a "Super Sidebar" serialization for the existing Sidebar
Panels. In order to prevent potential future mishaps, it is probably
best to generate both the "old" and the "new" sidebar structure from the
same underlying data:

1. It will be easier for us to ensure that _all_ current menu items will
   find a new home.
2. It allows us to iteratively group and/or rename menu items (or
   actively choose not rendering them in the future).

Currently we just default to the existing sidebar groupings, but the
current code is written in a way, that it would be easy for us to start
add a new "super sidebar parent" to the sidebar serialization, and e.g.
force each MenuItem to define one of those. Same goes e.g. with names
and other properties.