Fix 500 error for calls without JWT token to Git http
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/391771

**Problem**

GitHttpController is used by Workhorse to verify Git requests for HTTP
protocol. However, it's also possible to send a direct request to this
controller. But the controller is protected by JWT authentication. All
direct requests will raise a JWT token exception. This exception is not
handled and raises a 500 error.

**Solution**

Add a handler for JWT decode errors and return a Forbidden (403 error).

Changelog: fixed