Update documentation to match Praefect's new configuration structure
Omnibus preivously had a custom configuration structure for Praefect that
didn't match the actual structure of Praefect. This made a number of things
more difficult than they had to be. Omnibus was recently updated to place
Praefect's configuration under praefect['configuration'] hash which gets
directly marshaled as TOML for the final Praefect configuration. While there
is still backwards compatibility handling in place, we should update the
documentation to use the new structure in preparation for deprecating the
old configuration keys. This commit upgrades Praefect's documentation to
refer to the new keys.