Attempt to reinstall required Gitaly gems in tests
Before a test runs Gitaly, it attempts to run `bundle check` to ensure
all the required gems are installed. If they are not, the test aborts.

Instead of failing, let's try to recover via a `bundle install` and
only fail if the subsequent `bundle check` fails. This makes the tests
more resilient to users uninstalling gems or switching Ruby versions.