Include child pipeline scans in MR security widget
Fixes a bug whereby if security scans were run in a child pipeline
on the mainline branch, but are not run in a child pipeline on merge
request branches, previously detected findings show up as new in the
MR security widget.

This change ensures that any child pipeline reports are included when
fetching the security reports for a pipeline.

This MR also introduces some basic integration specs for the MR security
widget.