Add runner registration page
As part of the new registration workflow, this change adds a new page
where users can find all the required information to register their
runner.

In this change, page is empty while this feature is built.

Note that as part of this change, the token TTL was reduced from
3 hours to 1 hour.

Changelog: changed