Ensure top navbar exists when interacting with the DOM
This guards a couple places with the optional chaining operator to
ensure that the top navbar actually exists before trying to retrieve its
height or adding classes to it. This fixes an issue where a commits page
would be broken when the super sidebar is enabled. This also affects the
IDE, even though it currently does not support the super sidebar.