Simpler serialization for super sidebar
With https://gitlab.com/gitlab-org/gitlab/-/merge_requests/111186 we
introduced serialization methods for Panels and Menus, in order to be
able to use the existing Panels in the Navigation alpha.

In order to future-proof the concept, I decided to make those flexible,
by flattening the tree-structure of Menus and Menu Items and then
rebuilding it. The idea was to just provide different a different
`parent_id` to certain menu items and they would be zipped up
differently.

That being said, it is a little complicated and extending it to support
different groupings and different kind of menus (e.g. the static menu at
the top) proved difficult with this pattern. In a follow-up commit we
will introduce a new pattern which extends the existing Panels and
transforms them.