Add alternative Redis backend for Rack::Attack
This adds a feature-flagged new backend for Rack::Attack. The goal is
to reduce the number of Redis calls we make. To enable the new
backend, set the environment variable GITLAB_RACK_ATTACK_NEW_STORE=1.

If this works as expected on GitLab.com we will remove the old store
and the feature flag.

Changelog: other