Foundations for Super Sidebar categories
This is a new approach to Super Sidebar categories, while maintaining
the old menu data as the single-source of truth for menu items.

The idea behind this approach is simple:
1. Each menu item has a defined `super_sidebar_parent`. This will be the
   Menu the item is part of in the Super Sidebar
2. There is an optional `super_sidebar_before` field, if one needs to
   influence the order of the items.
3. The actual Panel for the Super Sidebar extends the existing Panel and
   mutates the internal "menu state". This transformation is done with a
   concern, in order to make this concept reusable for e.g. Groups.

In order to prove this concept, the existing Project menus for Issues,
Project Information and Wiki have been merged into the new plan menu.
All other menus have been copied untouched. The Develop and Operations
Menu Grouping can happen in a separate iteration, so that we can keep
the changes of this MR smaller.