Super Sidebar: Add uncategorized and removals
While we work on the conversions from old sidebar to Super Sidebar, it
is helpful to support both Uncategorized Menu Items and removals.

If we just drop all Menu Items which have no Super Sidebar defined, we
might miss to categorize items. Furthermore it enables us to move
certain items to a temporary category, while we work on categorizing
everything.

The explicit removal with the help of `::Sidebars::NilMenuItem` is
needed, because some menu entries are duplicated (e.g. Issues List
appears twice).