Release GitLab for Slack notifications
Removes the `integration_slack_app_notifications` from the codebase.

The feature is GitLab.com only, so has already been fully released when
the flag was toggled on prior, however, this commit removes all use of
the flag, and contains a commit trailer to add it to the changelog.

All documentation has already been merged.

Changelog: added
EE: true