Fix keyboard shortcuts modal in super sidebar
The help button was added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/109385, but it
wasn't working on certain pages, specifically those listed in
~/behaviors/shortcuts.js.

This removes the global function defined on pages other than those
listed above. Instead, we now rely on event delegation in the Shortcuts
class for detecting the clicks to the shortcuts modal trigger, since the
sidebar might not be in the DOM by the time the Shortcuts constructor
(or some subclass of it) is called.