Skip danger check on documentation MRs
While testing the release tooling for the maintenance policy, it was
noted documentation MRs fail with a Danger check. The check validates
merge requests targeting stable branches have the `type::bug` label, for
documentation changes it shouldn't be required to have the label because
doc updates are not bug fixes. This commit skips the danger check on
documentation merge requests.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2839