Fix inconsistent behaviour in design comments
The main issue is related to the use of the shortcut
ctrl/cmd + enter while saving the comment. Currently, the child
component DesignReplyForm is wrapped by ApolloMutation and then
used in the parent components(pages/index.vue, DesignNotes,
DesignDiscussion). Moreover, the autosave.js file binds the input
event to the text area. When the shortcut is executed, the
submitForm method emits submit-form to perform the mutation above
and parallely calls autosave.reset(). However, since the input is
bound in autosave.js, the save method is triggered again with a stale
value as the shortcut is also an input event. Handle the
comment value and the add/update mutations in the DesignReplyForm
component itself. The local storage value can be reset after component
is destroyed after checking if form is not getting submitted.

Changelog: fixed