Super sidebar: Render pill badges on menu items
Only caveat right now: If an item is both a section and has a pill, we
do not render the chevron.

While we currently have items which are both (e.g. the Issues menu in
the project), our designs do not have sections _with pills_.

So as soon as we refactored our grouping in the super sidebar
navigation, this should all work out properly.