Add cost settings which belongs to a runner
Add a cost settings table with the intent to store different
types of cost factors that apply for a specific shared runner.
Typically, we should expect the standard cost factor to vary
per runner. The other cost factors are applied based on the
attributes of the projects plan or wheather it is a contribution
to the gitlab open source project.

This just introduces the new columns which are not used yet
but will eventually be used in the Gitlab::Ci::Minutes::CostFactor
class.

Changelog: added
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/111977