Refactor protected environments list UI
GlAccordion doesn't allow for extra HTML in its header area, which we
need to show the rules counts. There were also some inconsistencies
between the designed UI and GlAccordion.

To save the footprint of the existing component, a new UI component was
created to iterate over the list of environments and pass it back via a
slot.