Reduce responsibility of #get_address_info
In some situations `UrlBlocker.validate!` considers a URI that we
cannot retrieve the `addrinfo` for to be valid.

This was previously signalled in code when `UrlBlocker#get_address_info`
would  return a `nil` value. The calling code in `.validate!` would then
handle the `nil` value in a way that treated the URI as valid.

This behaviour was reasonably opaque to the reader of the code.

This change makes this behaviour of `.validate!` more clear and
explicit.

See:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/112771#note_1292643144