Simplify execution of CreatePipelineService in rules spec
Defining a new method called #execute_service in the spec file, which
executes `described_class.new` in a separate parallel way to the more
typical `let` declaration is somewhat confusing. We've removed that
method and replaced the usage of it with a simple overriding `let`
block that calls the service class #execute method and passes in the
block that we want to use in that context.