Add notification when retrying trigger job
Behind the FF ci_recreate_downstream_pipeline, we allow use
to retry their trigger jobs. This has the effect of spawning
a new downstream pipeline. There are instances where while the
new pipeline is getting created, the old one is no longer visible
in the graph. For the MVC of this feature, showing a simple
toast message that explains that a downstream pipeline might
not yet be visible while being recreated.