Drop unnecessary storing of WebAuthn credential options in session
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26692 introduced
storing a serialized `WebAuthn::PublicKeyCredential::RequestOptions`
inside the session, but this appears to be unnecessary. The challenge
is properly stored in the session, and the credential options are
published to the frontend via Gon in JSON format.

This serialized structure caused problems in mixed deployments. With
the introduction of `WebAuthn::RelyingParty` in the `webauthn` 3.0
gem, `WebAuthn::PublicKeyCredential::RequestOptions` serializes its
`relying_party` data with this newly-introduced class. However, older
deployments do not have this class. As a result, sessions created in
this new format cannot be unmarshaled properly on older deployments.

Relates to https://gitlab.com/gitlab-com/gl-infra/production/-/issues/8481

Changelog: changed