Change order of CSS imports to fix search icon UI
The order of imports caused some styles to apply differently
from what is expected in GitLab UI (in particular, there are
styles in form_input that override styles in
search_box_by_type). This fix updates the order to be similar to
gitlab-ui's src/scss/components.scss and fixes the issue with
search icon being overlapped by the placeholder text due to
missing padding.

Changelog: fixed