Expose is_super_sidebar on Panel context
One might think: But hey, why would we need this, we could also add
certain transformations to the SuperSidebarPanel classes. But by
exposing this attribute in the context, it will be much simpler to e.g.
conditionally change the title or icon of a menu item.

My guideline would be:
1. Reordering / transformations of menu groups should be done with a
   SuperSidebar class that extends the existing panels
2. Simple adjustments like changing certain labels / icons should be
   done with `context.is_super_sidebar`