Add `release-environments` child pipeline setup and build images
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2674

In order to have commits to stable branches able to be deployed to
release environments for deployment testing and validation, we need
to build imags with the CNG project containing the contents of the
stable branches.

This change adds the work to get images built through the `CNG-mirror`
project for use by the release environments. These will only be built
on commits to a stable branch.