Rename read_ci_minute ability
- Rename the read_ci_minutes ability to admin_ci_minutes
because this permission will cover purchasing minutes and
viewing the usage quota page with a breakdown of how ci
minutes are being used. Customers who can read builds will
be allowed to see the total usage and quota summary but
not the full breakdown of the usage quota.
- Remove the admin_ci_minutes ability from project policy
since only group owners can see the usage quota breakdown page
as documented. The notifications where this ability is currently
used are only check the namespace anyways.