Prevent Ruby sessions from serializing arbitrary objects
The Rails and Rack session stores allow developers to store arbitrary
Ruby objects in the Hash, which gets serialized to Redis. However,
serializing objects may lead to multi-version incompatibilties because
there is no guarantee that the Ruby object is present in an older
version.

To safeguard against this problem, this patch checks that objects
stored in the session are in an allow list. Note that these checks are
restricted to test and development environments at the moment. Only
add to the allow list if you know that the object should be handled
gracefully in a mixed deployment.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/393554

Changelog: changed