Fail danger if package-and-test pipeline is canceled
As part of extending the maintenance policy, the package-and-test
pipeline execution should be enforced on stable branches. When this
pipeline is manual or canceled, Danger should fail indicating the MR
author an action is required from them.

This commit ensure Danger considers the downstream pipeline status,
rather than the bridge status.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2839