Prevent QA check on stable branches
Danger checks include a QA one to verify if a QA selector was modified,
and in this case, encourage the execution of the package-and-test
pipeline. Stable branches automatically trigger the package-and-test
pipeilne to guarantee changes are compliant from the Quality side.

Because of the above, this commit removes the QA check from stable
branches.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2839